changed return type of .count() from int to long + javadocing

+++ src/main/java/xawd/jflow/iterators/AbstractDoubleFlow.java
@@ -255,7 +255,7 @@ public abstract class AbstractDoubleFlow implements DoubleFlow
 	}
 
 	@Override
-	public int count()
+	public long count()
 	{
 		return DoubleReductionConsumption.count(this);
 	}
